requirements.txt
streamlit
pillow
PyMuPDF
transformers
torch
